Move GridBagHelper to its own class
